# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ちょっとレンタルサービスのAPI仕様書",
        "title": "Chotto Rental API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/users/me": {
            "put": {
                "description": "ログインユーザーの表示名を更新します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ユーザー"
                ],
                "summary": "ユーザー情報更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token} 形式",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "更新するユーザー情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_user.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新後のユーザー情報",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "リクエストが不正です",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "404": {
                        "description": "ユーザーが見つかりません",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "ユーザーと関連する全てのデータ（レビュー、マッチング、応募、案件）を削除します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ユーザー"
                ],
                "summary": "ユーザー削除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token} 形式",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "400": {
                        "description": "リクエストが不正です",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "ユーザーIDに基づいてユーザーの基本情報を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ユーザー"
                ],
                "summary": "ユーザー情報取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ユーザー情報",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "無効なユーザーID",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "404": {
                        "description": "ユーザーが見つかりません",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    }
                }
            }
        },
        "/users/{id}/reviews": {
            "get": {
                "description": "指定されたユーザーが受け取ったレビューの一覧を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ユーザー"
                ],
                "summary": "ユーザーのレビュー一覧取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "レビュー一覧",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.Review"
                            }
                        }
                    },
                    "400": {
                        "description": "無効なユーザーID",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_kageyama0_chotto-rental_internal_model.Case": {
            "description": "案件情報",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "durationMinutes": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "reward": {
                    "type": "integer"
                },
                "scheduledDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.User"
                },
                "userID": {
                    "description": "案件の依頼者",
                    "type": "string"
                }
            }
        },
        "github_com_kageyama0_chotto-rental_internal_model.Matching": {
            "description": "マッチング情報",
            "type": "object",
            "properties": {
                "arrivalConfirmationDeadline": {
                    "type": "string"
                },
                "arrivalConfirmedByHelper": {
                    "type": "boolean"
                },
                "arrivalConfirmedByRequester": {
                    "type": "boolean"
                },
                "case": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.Case"
                },
                "caseID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "helper": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.User"
                },
                "helperID": {
                    "description": "案件を受けたユーザー",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "meetingLocation": {
                    "type": "string"
                },
                "requester": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.User"
                },
                "requesterID": {
                    "description": "案件の依頼者",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_kageyama0_chotto-rental_internal_model.Review": {
            "description": "レビュー情報",
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "matching": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.Matching"
                },
                "matchingID": {
                    "type": "string"
                },
                "reviewedUser": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.User"
                },
                "reviewedUserID": {
                    "description": "レビューを書かれたユーザー",
                    "type": "string"
                },
                "reviewer": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.User"
                },
                "reviewerID": {
                    "description": "レビューを書いたユーザー",
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_kageyama0_chotto-rental_internal_model.User": {
            "description": "ユーザー情報",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "noShowCount": {
                    "type": "integer"
                },
                "passwordHash": {
                    "type": "string"
                },
                "trustScore": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_kageyama0_chotto-rental_pkg_util.Response": {
            "description": "HTTPレスポンス",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_handler_user.UpdateUserRequest": {
            "type": "object",
            "required": [
                "display_name"
            ],
            "properties": {
                "display_name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "「Bearer 」の後にJWTトークンを付与してください",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}